requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==20.1.0

